Get port id from the interface name.
In the CNI add API, the current implementation fetches
the port id of the interface from its name. It assumes the
name has the port id is part of the interface name as a
standard expression. This implementation is not reliable
in case the dpdk assigns a different port id for the interface.

Make it more dynamic and always fetch the port id by sending
a get request to the gNMI server.

Signed-off-by: Venkatesh Petla <neelakanta.venkatesh.petla@intel.com>