Add rule for host IP
Add a rule to the dpdk pipeline for forwarding the packets to/from host IP.
Forward them via the default host interface and its port id.

How to get the host IP from the inframanager pod?
Kubernetes populates the pod object in the apiserver with
spec:
nodeName: node1
status:
podIP: 172.168.0.2
hostIP: 192.168.0.1

Here hostIP is the nodeIP on which the pod is running on.

Inside the inframanager pod, read the node IP from the pod object and use it for
programming the forwarding rules.

Signed-off-by: Venkatesh Petla <neelakanta.venkatesh.petla@intel.com>